Be explicit about local tip caching
Anytime a player created or update a chaintree, it would never
reach back out to tupelo to get any newere tips. This resulted in
the player having stale trees when a remote actor changed some state.

This switches that caching to be explicit, enabled only for the playyer
and the home tree