Make debug & update on launch more robust
These were a little crashy before. Not totally sure why in the case of DEBUG. My theory is that some deeper library code was also looking for the old "DEBUG" var and slowing down and losing some sort of race condition. Ah, event-driven code. In the case of updating on launch, we're no longer racing to create the main window first, which is much better.